+++ android/app/src/main/java/com/example/budgetcircle/database/entities/main/OperationSum.kt
@@ -1,3 +1,3 @@
 package com.example.budgetcircle.database.entities.main
 
-data class OperationSum(val title: String, val sum: Float)
+data class OperationSum(val title: String, val sum: Double)
