+++ android/app/src/main/java/com/example/budgetcircle/database/entities/types/BudgetType.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity (tableName = "budget_types")
 data class BudgetType (
     var title: String,
-    var sum: Float,
+    var sum: Double,
     var isDeletable: Boolean
 ) {
     @PrimaryKey (autoGenerate = true)
